added edge routing avoiding node self-crossing
beta version (see test). \porthofinitewidth can use any anchor and
doesn't require specifying the direction to go. it used \portho (which
is a vertical-horizontal or horizontal-vertical decision maker),
however, there is significant redundancy in passing to that macro since
\porthofinitewidth already considers every possibility of where the
outport is on the boundary (NSEW), the inport is on the boundary (NSEW),
and the relative x (EW) and y (NS) between the outport and inport. it
would be better to include the drawing instructions under each branch,
though as the fact that portho works by rotation shows there is
significant repetition between branches.