testing for finite width edge routing
the finite width routing at least makes satisfactory joins between the
same nodes as the one that doesn't account for finite width. however, it
is unsatisfactory in two ways:

one, it only account for finite width in one dimension, and will
self-cross if the node is overbig in two dimensions because it just
passes the coordinates to build an S-join (as written, it works well for
tall and skinny or wide and short nodes, but not nodes which are big in
both dimensions relative to the distance between the other nodes). this
could be fixed by doing variable split S-join (currently it always uses
.5) based on finite node dimensions.

two, it incorrectly extends and then backtracks in at least one case.
this is due to the fact that the outport port may be, say, west of the
input port, but after extending past a boundary the effective port (used
for the middle L or S join) may be east of the input port, but the
outdir used by portho is still the original "go east". this can be
handled by evaluating what the difference in positions is after any
extensions.